大整数更新

+++ 数学/大整数/大整数取模.cpp
@@ -1,7 +1,7 @@
 //输出n%m的值，其中n<=10^10,m<=10^9
-int bigIntegerMod(const string&n,int m){
-    int ans=0;
-    for(int i=0;i<n.size();++i)
-        ans=(int)((ans*10ll+n[i]-'0')%m);
+long long bigIntegerMod(const string& n, long long m) {
+    long long ans = 0;
+    for (int i = 0; i < n.size(); ++i)
+        ans = ((ans * 10 + n[i] - '0') % m);
     return ans;
 }
